add css support, WIP selection editors
We're not managing state properly just yet, but we're getting close!
This brings in CSS support (which wasn't working, but untested up until
now). Typings are now generated for CSS Modules, which is great for
correctness. We're starting to render the first editors. Soon this will
be hooked up to the state management via hooks, which is the whole point
of this exercise/approach.