rename Selection to AppSelection
Selection is a built-in, always-imported type in TypeScript. Neglecting
to import Selection and shadowing the original Selection causes subtle
errors that are difficult to diagnose. Better to work around it by
renaming our Selection to AppSelection.

This commit might not be in working order. It's partially staged and I'd
rather not labor too much to make a working commit that I'm about to
commit over anyways.