fix exclusive selections
Exclusive selections now swaps the selections out. It no longer has all
of its selections present, yet disabled. This keeps our data storage
clean and accurately representing between both the reducer and the long
term storage (disk, perhaps). Additionally, it keeps the pure notion of
extant selections as intended. Now if we want something to appear all of
the time, we have fixed selections to help with that.