wip on order of battle
Orders of battle can now be created. Select a faction, and all units of
that faction can be added to the order of battle. Unbound selections are
not fully supported yet, and things that are not units are not supported
yet (though I think this is a data issue, and not a logic issue).

The army roster for the crusade forces aren't limited to the order of
battle yet - that's the next stop. That, and making those selections
finite.