wip towards finite pools
I ran into a lot of trouble with managing options via a vast and rigid
hierarchy. So I've flattened the options. Still a work in progress - the
data must be updated and the new functionality tested.

Flattening makes us more query based. We can query for certain kinds of
tags in the data to determine things like children selections to create
and hopefully pool options to get. Since we'll be fundamentally tag
driven, this should allow us some greater freedom with determining what
options are exactly available.